requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
python-environ==0.4.54
sqlparse==0.5.3
Unipath==1.1
